Rename interface to OkHttp3SocketClientFactory.java.  Modify ParseLiveQueryClient to accept OkHttpClient instances.
