+++ ParseLiveQuery/src/main/java/com/parse/ParseLiveQueryClient.java
@@ -3,8 +3,9 @@ package com.parse;
 import java.net.URI;
 import java.util.concurrent.Executor;
 
-public interface ParseLiveQueryClient {
+import okhttp3.OkHttpClient;
 
+public interface ParseLiveQueryClient {
     <T extends ParseObject> SubscriptionHandling<T> subscribe(ParseQuery<T> query);
 
     <T extends ParseObject> void unsubscribe(final ParseQuery<T> query);
@@ -23,6 +24,10 @@ public interface ParseLiveQueryClient {
 
         public static ParseLiveQueryClient getClient(URI uri) {
             return new ParseLiveQueryClientImpl(uri);
+        }
+
+        public static <T extends ParseObject> ParseLiveQueryClient getClient(URI uri, OkHttpClient okHttpClient) {
+            return new ParseLiveQueryClientImpl(uri, okHttpClient);
         }
 
         /* package */
